2-12-17 Dean
Built function in endpoint.php for editing the log.json by slicing the
string. This has to be done when not using strict .json files in PHP.
This fixes issue #4. Also built the function for rendering the log in
the table. This fixes issue #5. Some minor updates to the HTML and CSS
as well.